Initial commit
This is the initial version of the chatbot which will be designed to specialize in recommending movies. The chatbot uses OpenAI API to make the conversations and currently connects to a movie recommendation tool based on a simple keyword search (like "recommend", "suggest") and after that it makes recommendations based on a cosine similarity of movies in IMDB 1000 list. Both the recommendation algorithm and the logic to connect to recommendation from standard chatbot should be optimized in later versions but this is an initial skeleton also to make the GUI workable.